+++ src/store/reducers/queryResults.js
@@ -135,18 +135,20 @@ export function matchNewPostIntoQueryResults (state, { id, isPublic, type, group
   // Group streams
   return reduce((memo, group) => {
     queriesToMatch.push(
-      { context: 'groups', slug: group.slug },
+      { context: 'groups', slug: group.slug, activePostsOnly: false },
+      { context: 'groups', slug: group.slug, activePostsOnly: true }, // For custom views
       { context: 'groups', slug: group.slug, groupSlugs: [group.slug] }, // For FETCH_POSTS_MAP
-      { context: 'groups', slug: group.slug, filter: type },
-      { context: 'groups', slug: group.slug, sortBy: 'updated' },
+      { context: 'groups', slug: group.slug, filter: type, activePostsOnly: false },
+      { context: 'groups', slug: group.slug, sortBy: 'updated', activePostsOnly: false },
+      { context: 'groups', slug: group.slug, sortBy: 'updated', activePostsOnly: true }, // For custom views
       { context: 'groups', slug: group.slug, sortBy: 'updated', search: '', groupSlugs: [group.slug] }, // For FETCH_POSTS_MAP_DRAWER
-      { context: 'groups', slug: group.slug, sortBy: 'updated', filter: type },
-      { context: 'groups', slug: group.slug, sortBy: 'created' },
+      { context: 'groups', slug: group.slug, sortBy: 'updated', filter: type, activePostsOnly: false },
+      { context: 'groups', slug: group.slug, sortBy: 'created', activePostsOnly: false },
+      { context: 'groups', slug: group.slug, sortBy: 'created', activePostsOnly: true }, // For custom views
       { context: 'groups', slug: group.slug, sortBy: 'created', search: '', groupSlugs: [group.slug] }, // For FETCH_POSTS_MAP_DRAWER
-      { context: 'groups', slug: group.slug, sortBy: 'created', filter: type },
-      // For events stream
+      { context: 'groups', slug: group.slug, sortBy: 'created', filter: type, activePostsOnly: false },
+      // For events stream upcoming events
       { context: 'groups', slug: group.slug, sortBy: 'start_time', filter: type, order: 'asc' },
-      { context: 'groups', slug: group.slug, sortBy: 'start_time', filter: type, order: 'desc' }
     )
     for (let topic of topics) {
       queriesToMatch.push(