+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to Hylo Evo (the Hylo front-end) will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Newly created posts appear immediately in stream, as well as projects view and upcoming events view if appropriate
+
 ## [5.0.4] - 2022-11-23
 
 ### Fixed